+++ Modules/Intent.Modules.Integration.HttpClients.Shared/NuGetPackages.cs
@@ -1,19 +1,21 @@
 using Intent.Engine;
+using Intent.Modules.Common.CSharp.Nuget;
 using Intent.Modules.Common.VisualStudio;
 
 namespace Intent.Modules.Integration.HttpClients.Shared;
 
 public static class NuGetPackages
 {
-    public static INugetPackageInfo MicrosoftExtensionsHttp(IOutputTarget outputTarget) => new NugetPackageInfo(
-        name: "Microsoft.Extensions.Http",
-        version: outputTarget.GetMaxNetAppVersion() switch
-        {
-            (5, 0) => "5.0.0",
-            (6, 0) => "6.0.0",
-            (7, 0) => "7.0.0",
-            _ => "8.0.0"
-        });
-    public static readonly INugetPackageInfo MicrosoftAspNetCoreWebUtilities = new NugetPackageInfo("Microsoft.AspNetCore.WebUtilities", "2.2.0");
-    public static readonly INugetPackageInfo IdentityModelAspNetCore = new NugetPackageInfo("IdentityModel.AspNetCore", "4.3.0");
+    public const string IdentityModelAspNetCorePackageName = "IdentityModel.AspNetCore";
+    public const string MicrosoftAspNetCoreWebUtilitiesPackageName = "Microsoft.AspNetCore.WebUtilities";
+    public const string MicrosoftExtensionsHttpPackageName = "Microsoft.Extensions.Http";
+    public const string SystemTextJsonPackageName = "System.Text.Json";
+
+    public static NugetPackageInfo IdentityModelAspNetCore(IOutputTarget outputTarget) => NugetRegistry.GetVersion(IdentityModelAspNetCorePackageName, outputTarget.GetMaxNetAppVersion());
+
+    public static NugetPackageInfo MicrosoftAspNetCoreWebUtilities(IOutputTarget outputTarget) => NugetRegistry.GetVersion(MicrosoftAspNetCoreWebUtilitiesPackageName, outputTarget.GetMaxNetAppVersion());
+
+    public static NugetPackageInfo MicrosoftExtensionsHttp(IOutputTarget outputTarget) => NugetRegistry.GetVersion(MicrosoftExtensionsHttpPackageName, outputTarget.GetMaxNetAppVersion());
+
+    public static NugetPackageInfo SystemTextJson(IOutputTarget outputTarget) => NugetRegistry.GetVersion(SystemTextJsonPackageName, outputTarget.GetMaxNetAppVersion());
 }